Fix type of the cardinality argument

+++ sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/ExtractBenchmark.scala
@@ -48,7 +48,7 @@ object ExtractBenchmark extends SqlBasedBenchmark {
     }
   }
 
-  private def run(cardinality: Int, field: String): Unit = {
+  private def run(cardinality: Long, field: String): Unit = {
     codegenBenchmark(s"$field of timestamp", cardinality) {
       doBenchmark(cardinality, s"EXTRACT($field FROM (cast(id as timestamp)))")
     }
